+++ src/main/java/dev/yatloaf/modkrowd/cubekrowd/message/cache/MissileWarsMessageCache.java
@@ -1,12 +1,15 @@
 package dev.yatloaf.modkrowd.cubekrowd.message.cache;
 
 import dev.yatloaf.modkrowd.cubekrowd.message.Message;
+import dev.yatloaf.modkrowd.cubekrowd.message.MissileWarsDeathMessage;
 import dev.yatloaf.modkrowd.cubekrowd.message.MissileWarsGameEndMessage;
 import dev.yatloaf.modkrowd.cubekrowd.common.cache.TextCache;
 import dev.yatloaf.modkrowd.cubekrowd.subserver.MissileWarsSubserver;
 import dev.yatloaf.modkrowd.util.text.StyledStringReader;
+import net.minecraft.text.TranslatableTextContent;
 
 public class MissileWarsMessageCache extends MinigameMessageCache {
+    private MissileWarsDeathMessage missileWarsDeathMessage;
     private MissileWarsGameEndMessage missileWarsGameEndMessage;
 
     public MissileWarsMessageCache(TextCache original, MissileWarsSubserver subserver) {
@@ -15,12 +18,28 @@ public class MissileWarsMessageCache extends MinigameMessageCache {
 
     @Override
     protected Message createResult() {
+        MissileWarsDeathMessage missileWarsDeathMessage = this.missileWarsDeathMessageFast();
+        if (missileWarsDeathMessage.isReal()) {
+            return missileWarsDeathMessage;
+        }
+
         MissileWarsGameEndMessage missileWarsGameEndMessage = this.missileWarsGameEndMessageFast();
         if (missileWarsGameEndMessage.isReal()) {
             return missileWarsGameEndMessage;
         }
 
         return super.createResult();
+    }
+
+    public final MissileWarsDeathMessage missileWarsDeathMessageFast() {
+        if (this.missileWarsDeathMessage == null) {
+            if (this.original.text().getContent() instanceof TranslatableTextContent content) {
+                this.missileWarsDeathMessage = MissileWarsDeathMessage.parseFast(content);
+            } else {
+                this.missileWarsDeathMessage = MissileWarsDeathMessage.FAILURE;
+            }
+        }
+        return this.missileWarsDeathMessage;
     }
 
     public final MissileWarsGameEndMessage missileWarsGameEndMessageFast() {