+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Tie Detector feature.
 - Indicators for custom messages added by the mod.
+- MissileWars death message theming.
 
 ### Changed